rename core src files to match ts convention
